+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project are documented here. The format is based on
 
 ## Unreleased
 
+-   Aligned to Luxon 0.0.6.
 -   Added JavaScript Date <-> DateTime pipes.
 
 ## 0.0.1 (2017-10-13)